requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.4
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
Babel==2.9.1
blinker==1.4
cachetools==4.2.4
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.6
click==6.7
cloud-init==22.2
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
defusedxml==0.7.1
diff-match-patch==20200713
distro-info===0.18ubuntu0.18.04.1
Django==3.2.2
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.13.0
django-filter==21.1
django-import-export==2.5.0
django-jazzmin==2.5.0
django-jsonfield==1.4.1
django-money==2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
et-xmlfile==1.1.0
future==0.18.2
google-api-core==2.2.1
google-api-python-client==2.28.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
html5lib==1.1
httplib2==0.20.2
hyperlink==17.3.1
idna==3.2
incremental==16.10.1
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
macaroonbakery==1.1.3
MarkupPy==1.14
MarkupSafe==1.0
netifaces==0.10.4
numpy==1.20.3
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.7
PAM==0.4.2
pandas==1.2.4
pbr==5.8.1
Pillow==8.2.0
protobuf==3.19.1
py-moneyed==1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pyfcm==1.5.4
PyGObject==3.26.1
PyJWT==2.4.0
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.1.2
pyOpenSSL==17.5.0
pyparsing==3.0.4
PyPDF2==1.26.0
pyRFC3339==1.0
pyserial==3.4
python-apt==1.6.5+ubuntu0.7
python-barcode==0.13.1
python-bidi==0.4.2
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2021.1
pyxdg==0.25
PyYAML==5.4.1
razorpay==1.2.0
reportlab==3.5.67
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
rsa==4.7.2
SecretStorage==2.3.1
service-identity==16.0.0
six==1.16.0
sqlparse==0.4.1
ssh-import-id==5.7
stevedore==3.5.0
system-service==0.3
systemd-python==234
tablib==3.0.0
Twisted==17.9.0
typing-extensions==3.10.0.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==15.1.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
webencodings==0.5.1
xhtml2pdf==0.2.5
xkit==0.0.0
xlrd==2.0.1
xlwt==1.3.0
zope.interface==4.3.2
